refactor: simplify comments and move the new refund function to offline-send to align with the current semantic
