+++ core/api/src/app/payments/send-lightning.ts
@@ -1007,13 +1007,8 @@ const lockedPaymentViaLnSteps = async ({
       pubkey: outgoingNodePubkey,
     })
   } else {
-    // Model 2: btcProtocolAndBankFee is the accounting TOTAL (routing reserve +
-    // service fee). The LND routing budget / verifyMaxFee 0.5% cap must see the
-    // routing reserve ONLY — recover it as total − bankFee, never the total, or
-    // the cap trips and probing can consume the service fee.
+    // btcProtocolAndBankFee is the TOTAL (routing reserve + service fee)
     const maxFeeCheckArgs = {
-      // Clamp to ≥ ZERO as defense against a corrupt recovered bank fee >
-      // total; the live path can never underflow.
       maxFeeAmount: calc.max(
         calc.sub(paymentFlow.btcProtocolAndBankFee, paymentFlow.btcBankFee),
         ZERO_SATS,
